requirements.txt
streamlit
redis
faiss-cpu # Or faiss-gpu if you have CUDA installed
requests # For potential API calls (like SiliconFlow)
python-dotenv # To load environment variables from .env
numpy # For numerical operations, especially with FAISS/embeddings
beautifulsoup4 # For web scraping in the crawler module
